Add Morse monitor UI and display system
Introduces Morse monitor UI components, object constructors, and a display system to visualize Morse code input timing and regions. Updates rendering and vector utilities to support new UI features, and adjusts Morse code timing and error margin constants. Also refines artillery control and receiver systems for improved message handling and region validation.

+++ include/systems/core/rendering/rectangle_render_system.h
@@ -15,7 +15,7 @@ public:
 
 	static void DrawUi(
 		entt::registry& registry, 
-		Nc::Vector2i windowSize
+		Nc::Vector2f windowSize
 	);
 
 };
